docker-api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
PyYAML==6.0.2
opendataloader-pdf
requests==2.32.3
